+++ src/components/card/PostContent.jsx
@@ -1,0 +1,15 @@
+import { PostCont, PostDate } from './PostStyle';
+
+const PostContent = () => {
+  return (
+    <>
+      <PostCont>
+        포토레시피가 추천하는 기억에 남을 크리스마스를 위한 근사한 포토 스팟을
+        알려드립니다🎄 이 곳만은 꼭! 가서 멋진 사진을 찍고 공유해주세요🎅
+      </PostCont>
+      <PostDate>2022년 12월 13일</PostDate>
+    </>
+  );
+};
+
+export default PostContent;
